refactor(repo): move appropriate constants to constants/config.ts
