--- constants/columns.ts
@@ -1,1 +1,0 @@
-export const initialColumns = ['non-planned', 'todo', 'doing', 'done']
